+++ pytorch_lightning/trainer/connectors/accelerator_connector.py
@@ -849,13 +849,5 @@ class AcceleratorConnector:
         return is_distributed
 
     @property
-    def has_ipu(self) -> bool:
-        return isinstance(self.accelerator, IPUAccelerator) and isinstance(self.strategy, IPUStrategy)
-
-    @property
-    def use_ipu(self) -> bool:
-        return isinstance(self.accelerator, IPUAccelerator)
-
-    @property
     def use_dp(self) -> bool:
         return isinstance(self.strategy, DataParallelStrategy)